+++ src/index.html
@@ -13,15 +13,14 @@
     />
   </head>
   <body>
+    <include src="./partials/loader.html"></include>
     <include src="./partials/header.html"></include>
     <include src="./partials/films.html"></include>
     <include src="./partials/footer.html"></include>
-    <include src="./partials/ModalCard.html"></include>
+    <include src="./partials/modalCard.html"></include>
 
     <include src="./partials/myTeam.html"></include>
 
     <script type="module" src="./js/index.js"></script>
-    
   </body>
-
 </html>
