--- src/js/apiFilms/apiKey.js
@@ -1,1 +1,0 @@
-export const API_KEY = 'cbb822aa772537b57d5d040000698149'
