--- src/js/apiFilms/baseUrl.js
@@ -1,1 +1,0 @@
-export const BASE_URL = 'https://api.themoviedb.org/3';
