+++ src/js/index.js
@@ -1,6 +1,7 @@
-import './spinner'
-import './films';
-import './apiFilms/apiSearch';
+import './loader';
+import './fetchMoviesAPIService';
+import './onHomePageLoad';
+import './movieSearchService';
+import './infiniteScroll';
 import './modal';
-// import './pagination';
 import './modal-team';
