+++ src/js/loader.js
@@ -1,0 +1,16 @@
+// import LoadMoreBtn from './load-more-btn'
+
+// const container = document.querySelector('body')
+
+// window.addEventListener('load', function() {
+//     displayLoader();
+// });
+
+// export function displayLoader() {
+//     container.insertAdjacentHTML('beforeBegin', '<div class="loader"></div>');
+// }
+
+// const loadMoreBtn = new LoadMoreBtn({
+//     selector: '[data-action="load-more"]',
+//     hidden: true,
+// });
