+++ src/js/refs.js
@@ -2,6 +2,11 @@ function getRefs() {
   return {
     filmGallery: document.querySelector('.film-gallery'),
     guard: document.querySelector('.guard-js'),
+    searchForm: document.querySelector('.js-search__form'),
+    watchedBtn: document.querySelector('.header_btn-watched'),
+    queueBtn: document.querySelector('.header_btn-queue'),
+    backdrop: document.querySelector('.backdrop'),
+    closeButton: document.querySelector('.close-button'),
   };
 }
