+++ src/partials/films.html
@@ -1,7 +1,7 @@
-
-<section class="container film-contrainer">
+<main>
+  <section class="container film-contrainer">
     <h2 class="visually-hidden">Films</h2>
-    <ul class="film-gallery">
-    </ul>
+    <ul class="film-gallery"></ul>
     <div class="guard-js"></div>
-</section>
+  </section>
+</main>
