+++ src/partials/loader.html
@@ -1,1 +1,14 @@
-<div class="loader"></div>
+<div class="preloader">
+  <div class="spinner-container">
+    <div class="spinner">
+      <div class="spinner__item">
+        <div class="spinner__circle"></div>
+      </div>
+      <div class="spinner__item">
+        <div class="spinner__circle"></div>
+      </div>
+    </div>
+  </div>
+</div>
+
+<!-- <div class="loader"></div> -->
