--- src/partials/spiner.html
@@ -1,3 +1,0 @@
-<span class="spinner-border spinner-border-sm spinner is-hidden " role="status" aria-hidden="true"></span>
-<span class="label"></span>
-
